рефакторинг попапа
