--- src/components/film-details.js
@@ -1,144 +1,0 @@
-// Шаблон Подробная информация о фильме (попап)
-import {MONTHS} from "../const.js";
-import AbstractComponent from "./abstract-component.js";
-
-const createGenresTemplate = (genre) => {
-
-  return (
-    ` <span class="film-details__genre">${genre}</span>`
-  );
-};
-
-const createFilmInfoPopupTemplate = (film) => {
-  const {name,
-    poster,
-    rating,
-    age,
-    director,
-    writers,
-    actors,
-    releaseDate,
-    duration,
-    country,
-    genre,
-    description
-  } = film;
-  const genresTemplate = genre.map((it) => createGenresTemplate(it)).join(`\n`);
-  const filmDate = `${releaseDate.getDate()} ${MONTHS[releaseDate.getMonth()]} ${releaseDate.getFullYear()}`;
-
-  return (
-    `<section class="film-details">
-      <form class="film-details__inner" action="" method="get">
-        <div class="form-details__top-container">
-          <div class="film-details__close">
-            <button class="film-details__close-btn" type="button">close</button>
-          </div>
-          <div class="film-details__info-wrap">
-            <div class="film-details__poster">
-              <img class="film-details__poster-img" src="${poster}" alt="">
-    
-              <p class="film-details__age">${age}+</p>
-            </div>
-    
-            <div class="film-details__info">
-              <div class="film-details__info-head">
-                <div class="film-details__title-wrap">
-                  <h3 class="film-details__title">${name}</h3>
-                  <p class="film-details__title-original">Original: ${name}</p>
-                </div>
-    
-                <div class="film-details__rating">
-                  <p class="film-details__total-rating">${rating}</p>
-                </div>
-              </div>
-    
-              <table class="film-details__table">
-                <tr class="film-details__row">
-                  <td class="film-details__term">Director</td>
-                  <td class="film-details__cell">${director}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Writers</td>
-                  <td class="film-details__cell">${writers}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Actors</td>
-                  <td class="film-details__cell">${actors}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Release Date</td>
-                  <td class="film-details__cell">${filmDate}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Runtime</td>
-                  <td class="film-details__cell">${duration}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">Country</td>
-                  <td class="film-details__cell">${country}</td>
-                </tr>
-                <tr class="film-details__row">
-                  <td class="film-details__term">${genre.length > 1 ? `Genres` : `Genre`}</td>
-                  <td class="film-details__cell">
-                    ${genresTemplate}
-                  </td>
-                </tr>
-              </table>
-    
-              <p class="film-details__film-description">
-                ${description}
-              </p>
-            </div>
-          </div>
-    
-          <section class="film-details__controls">
-            <input type="checkbox" class="film-details__control-input visually-hidden" id="watchlist" name="watchlist">
-            <label for="watchlist" class="film-details__control-label film-details__control-label--watchlist">Add to watchlist</label>
-    
-            <input type="checkbox" class="film-details__control-input visually-hidden" id="watched" name="watched">
-            <label for="watched" class="film-details__control-label film-details__control-label--watched">Already watched</label>
-    
-            <input type="checkbox" class="film-details__control-input visually-hidden" id="favorite" name="favorite">
-            <label for="favorite" class="film-details__control-label film-details__control-label--favorite">Add to favorites</label>
-          </section>
-        </div>
-      </form>
-    </section>`
-  );
-};
-
-export default class FilmInfoPopup extends AbstractComponent {
-  constructor(film) {
-    super();
-
-    this._film = film;
-    this._PopupButtonClose = this.getElement().querySelector(`.film-details__close-btn`);
-  }
-
-  getTemplate() {
-    return createFilmInfoPopupTemplate(this._film);
-  }
-
-  setPopupButtonCloseHandler(handler) {
-    this._PopupButtonClose.addEventListener(`click`, handler);
-  }
-
-  removeElement(handler) {
-    this._PopupButtonClose.addEventListener(`click`, handler);
-  }
-
-  setWatchlistButtonClickHandler(handler) {
-    this.getElement().querySelector(`.film-details__control-label--watchlist`)
-      .addEventListener(`click`, handler);
-  }
-
-  setWatchedButtonClickHandler(handler) {
-    this.getElement().querySelector(`.film-details__control-label--watched`)
-      .addEventListener(`click`, handler);
-  }
-
-  setFavoriteButtonClickHandler(handler) {
-    this.getElement().querySelector(`.film-details__control-label--favorite`)
-      .addEventListener(`click`, handler);
-  }
-}
