+++ src/controllers/movie-controller.js
@@ -1,10 +1,7 @@
 import FilmCardComponent from "../components/film-card.js";
-import FilmInfoPopupComponent from "../components/film-details.js";
-import FilmCommentsPopupComponent from "../components/film-comments.js";
-import {generateComments} from "../mock/comment.js";
-import {render, replace, addPopupElement, removePopupElement, RenderPosition} from "../utils/render.js";
+import FilmPopupComponent from "../components/film-popup.js";
+import {render, replace, remove, RenderPosition} from "../utils/render.js";
 
-const comments = generateComments();
 const siteBodyElement = document.querySelector(`body`);
 
 export default class MovieController {
@@ -13,7 +10,7 @@ export default class MovieController {
     this._onDataChange = onDataChange;
 
     this._filmCardComponent = null;
-    this._filmInfoPopupComponent = null;
+    this._filmPopupComponent = null;
 
     this._onEscKeyDown = this._onEscKeyDown.bind(this);
   }
@@ -21,23 +18,21 @@ export default class MovieController {
   render(film) {
     // Рэндеринг карточек фильмов
     const oldFilmCardComponent = this._filmCardComponent;
-    const oldFilmInfoPopupComponent = this._filmInfoPopupComponent;
-    const oldFilmPopupCommentComponent = this._filmPopupCommentComponent;
+    const oldFilmPopupComponent = this._filmPopupComponent;
 
     this._filmCardComponent = new FilmCardComponent(film);
-    this._filmInfoPopupComponent = new FilmInfoPopupComponent(film);
-    this._filmPopupCommentComponent = new FilmCommentsPopupComponent(comments);
+    this._filmPopupComponent = new FilmPopupComponent(film);
 
     const onButtonPopupClose = () => {
       this._removePopup(onButtonPopupClose);
     };
 
     this._filmCardComponent.setDetailClickHandler(() => {
-      addPopupElement(siteBodyElement, this._filmInfoPopupComponent, this._filmPopupCommentComponent);
+      render(siteBodyElement, this._filmPopupComponent, RenderPosition.BEFOREEND);
       document.addEventListener(`keydown`, this._onEscKeyDown);
     });
 
-    this._filmInfoPopupComponent.setPopupButtonCloseHandler(onButtonPopupClose);
+    this._filmPopupComponent.setPopupButtonCloseHandler(onButtonPopupClose);
 
     // Обработка кликов на кнопках   карточек фильмов «Add to watchlist», «Already watched», «Add to favorites»
     this._filmCardComponent.setWatchlistButtonClickHandler((evt) => {
@@ -62,39 +57,38 @@ export default class MovieController {
     });
 
     // Обработка кликов на кнопках попапа «Add to watchlist», «Already watched», «Add to favorites»
-    this._filmInfoPopupComponent.setWatchlistButtonClickHandler((evt) => {
+    this._filmPopupComponent.setWatchlistButtonClickHandler((evt) => {
       evt.preventDefault();
       this._onDataChange(this, film, Object.assign({}, film, {
         isInWatchlist: !film.isInWatchlist,
       }));
     });
 
-    this._filmInfoPopupComponent.setWatchedButtonClickHandler((evt) => {
+    this._filmPopupComponent.setWatchedButtonClickHandler((evt) => {
       evt.preventDefault();
       this._onDataChange(this, film, Object.assign({}, film, {
         isInHistory: !film.isInHistory,
       }));
     });
 
-    this._filmInfoPopupComponent.setFavoriteButtonClickHandler((evt) => {
+    this._filmPopupComponent.setFavoriteButtonClickHandler((evt) => {
       evt.preventDefault();
       this._onDataChange(this, film, Object.assign({}, film, {
         isFavorit: !film.isFavorit,
       }));
     });
 
-    if (oldFilmCardComponent && oldFilmInfoPopupComponent && oldFilmPopupCommentComponent) {
+    if (oldFilmCardComponent && oldFilmPopupComponent) {
       replace(this._filmCardComponent, oldFilmCardComponent);
-      replace(this._filmInfoPopupComponent, oldFilmInfoPopupComponent);
-      replace(this._filmPopupCommentComponent, oldFilmPopupCommentComponent);
+      replace(this._filmPopupComponent, oldFilmPopupComponent);
     } else {
       render(this._container, this._filmCardComponent, RenderPosition.BEFOREEND);
     }
   }
 
   _removePopup(element) {
-    this._filmInfoPopupComponent.removeElement(element);
-    removePopupElement(siteBodyElement, this._filmInfoPopupComponent, this._filmPopupCommentComponent);
+    this._filmPopupComponent.removeElement(element);
+    remove(this._filmPopupComponent);
   }
 
   _onEscKeyDown(evt) {
